Register "Extra->Device"; break ExtraMenus -> DeviceToolBar
